Added signals and slots for the sequence displayer
